refactor: Move FakeViewingScopeValidator to fixtures
