+++ nvi-commons/src/testFixtures/java/no/sikt/nva/nvi/common/validator/FakeViewingScopeValidator.java
@@ -1,8 +1,7 @@
-package no.sikt.nva.nvi.test;
+package no.sikt.nva.nvi.common.validator;
 
 import java.net.URI;
 import java.util.List;
-import no.sikt.nva.nvi.common.validator.ViewingScopeValidator;
 
 public class FakeViewingScopeValidator implements ViewingScopeValidator {
